remove chunked paged prefill. expose arg flag
