+++ vllm/envs.py
@@ -57,7 +57,6 @@ if TYPE_CHECKING:
     VLLM_XLA_CACHE_PATH: str = os.path.join(VLLM_CACHE_ROOT, "xla_cache")
     VLLM_XLA_CHECK_RECOMPILATION: bool = False
     VLLM_SPARSE_INDEXER_MAX_LOGITS_MB: int = 512
-    VLLM_SPARSE_INDEXER_DISABLE_PAGED_PREFILL_CHUNKING: bool = False
     VLLM_USE_RAY_COMPILED_DAG_CHANNEL_TYPE: Literal["auto", "nccl", "shm"] = "auto"
     VLLM_USE_RAY_COMPILED_DAG_OVERLAP_COMM: bool = False
     VLLM_USE_RAY_WRAPPED_PP_COMM: bool = True
@@ -895,12 +894,6 @@ environment_variables: dict[str, Callable[[], Any]] = {
     # Default: 512 MB
     "VLLM_SPARSE_INDEXER_MAX_LOGITS_MB": lambda: int(
         os.getenv("VLLM_SPARSE_INDEXER_MAX_LOGITS_MB", "512")
-    ),
-    # If set, SM100 sparse-indexer paged prefill uses one unchunked DeepGEMM
-    # metadata batch with a PyTorch-built schedule. This is only for
-    # experiments because materialized paged-prefill logits can be very large.
-    "VLLM_SPARSE_INDEXER_DISABLE_PAGED_PREFILL_CHUNKING": lambda: bool(
-        int(os.getenv("VLLM_SPARSE_INDEXER_DISABLE_PAGED_PREFILL_CHUNKING", "0"))
     ),
     # If set, the OpenAI API server will stay alive even after the underlying
     # AsyncLLMEngine errors and stops serving requests
